+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "birbs",
-  "version": "0.5.5",
+  "version": "0.5.15",
   "description": "Capture and admire all your joyful events with this event manager and encapsulator!",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
