+++ src/manager.ts
